third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-color-space-list-expected.txt

This is a testharness.js-based test.
PASS Animating a custom property of type <color>+
PASS Animating a custom property of type <color>+ with a single keyframe
PASS Animating a custom property of type <color>+ with additivity
PASS Animating a custom property of type <color>+ with a single keyframe and additivity
FAIL Animating a custom property of type <color>+ with iterationComposite assert_equals: expected "rgb(125, 125, 125) rgb(250, 250, 250)" but got "rgb(25, 25, 25) rgb(50, 50, 50)"
PASS Animating a custom property of type <color>+ with different lengths is discrete
Harness: the test ran to completion.
